RBAC基于角色的存取控制: 1. 配置rbac 2. 建立授权数据 3. 执行权限检查

+++ backend/controllers/PostController.php
@@ -7,6 +7,7 @@ use common\models\Post;
 use common\models\PostSearch;
 use yii\filters\AccessControl;
 use yii\web\Controller;
+use yii\web\ForbiddenHttpException;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 
@@ -83,6 +84,10 @@ class PostController extends Controller
      */
     public function actionCreate()
     {
+        if (!Yii::$app->user->can('createPost')) {
+            throw new ForbiddenHttpException('对不起，你没有进行该操作的权限。');
+        }
+
         $model = new Post();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -103,6 +108,10 @@ class PostController extends Controller
      */
     public function actionUpdate($id)
     {
+        if (!Yii::$app->user->can('updatePost')) {
+            throw new ForbiddenHttpException('对不起，你没有进行该操作的权限。');
+        }
+
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -123,6 +132,10 @@ class PostController extends Controller
      */
     public function actionDelete($id)
     {
+        if (!Yii::$app->user->can('deletePost')) {
+            throw new ForbiddenHttpException('对不起，你没有进行该操作的权限。');
+        }
+
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
